+++ tests/multicolor_test.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+from object_colors import Color
+
+
+class TestMultiColor:
+    _ = Color()